refactor: support torch like separated modules
