+++ slowtorch/nn/__init__.py
@@ -1,10 +1,10 @@
 """\
-SlowTorch Neural Network APIs
-=============================
+SlowTorch Neural Network
+========================
 
 Author: Akshay Mestry <xa@mes3.dev>
 Created on: Wednesday, January 15 2025
-Last updated on: Friday, January 16 2025
+Last updated on: Sunday, June 01 2025
 """
 
 from . import functional
